requirements.txt
argcomplete==1.10.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.11.5
asttokens==2.0.5
attrs==21.4.0
awscli==1.25.26
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.8.2
bleach==5.0.1
blis==0.7.8
boto==2.49.0
boto3==1.24.26
botocore==1.27.26
bpyutils==0.5.2
bs4==0.0.1
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.5.18.1
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
compressed-rtf==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cymem==2.0.6
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.5.1
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.2.2
django-extensions==3.1.5
django-filter==21.1
django-haystack==3.2.1
django-import-export==2.8.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docutils==0.16
docx2txt==0.8
ebcdic==1.1.1
elastic-transport==8.1.2
elasticsearch==8.2.2
elasticsearch-dsl==7.4.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
extract-msg==0.28.7
fabric==2.7.0
fastjsonschema==2.16.1
filelock==3.7.1
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
gspread==5.4.0
html5lib==1.1
httpie==3.2.1
httplib2==0.20.4
idna==3.3
IMAPClient==2.1.0
invoke==1.7.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
langcodes==3.3.0
lazy-object-proxy==1.7.1
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
multidict==6.0.2
murmurhash==1.0.7
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
odfpy==1.4.1
olefile==0.46
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pathlib2==2.3.7.post1
pathy==0.6.2
pdfminer.six==20191110
pdfplumber==0.7.1
pickleshare==0.7.5
Pillow==9.1.1
pipupgrade==1.12.0
platformdirs==2.5.2
preshed==3.0.6
prettytable==3.3.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.8.2
Pygments==2.12.0
PyJWT==2.4.0
pylint==2.14.1
PyNaCl==1.5.0
pyparsing==3.0.9
PyPDF2==2.4.2
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.6
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.7.9
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.4.1
rsa==4.7.2
s3transfer==0.6.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.12.0
smart-open==5.2.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
SpeechRecognition==3.8.1
sqlparse==0.4.2
srsly==2.4.3
stack-data==0.3.0
tablib==3.2.1
terminado==0.15.0
textract==1.6.5
thinc==8.0.17
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
Wand==0.6.7
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
Whoosh==2.7.4
widgetsnbextension==3.6.1
wrapt==1.14.1
xlrd==1.2.0
XlsxWriter==3.0.3
xlwt==1.3.0
